requirements.txt
fastapi
uvicorn
openai
pinecone-client
python-multipart
requests
PyMuPDF
python-dotenv
